Move VM sizeof helpers into core
